requirements.txt
tensorflow==2.16.1
keras
keras-facenet==0.3.2
numpy==1.26.4
scipy==1.10.1
mtcnn==0.1.0
opencv-python==4.5.5.64
scikit-learn==1.3.2
chromadb==0.4.24
pandas==1.5.3
pillow

